fix(api): fix project role and style

fix project role and style

See merge request !771

+++ engine-ui/packages/console-ui/src/webapps/dataApi/views/projectManage/role/index.tsx
@@ -44,7 +44,7 @@ class RoleManagement extends React.Component<any, any> {
     }
 
     removeRole = (role: any) => {
-        Api.deleteRole({ roleId: role.id }).then((res: any) => {
+        Api.deleteRole({ roleId: role && role.id }).then((res: any) => {
             if (res.code === 1) {
                 message.success('移除角色成功！')
                 this.loadRoles()
